Ensure graceful shutdown on time out while waiting for birth dependencies (#5)
When a time out happens while waiting for birth dependencies, an error occurs while running the `fatalf` function to gracefully shut down the application with an non-zero exit code. This issue is caused by closing a channel that is `nil` at time of exection. This channel is only set when the child proces is actually started, which is not the case if we reach the state in which we time out while waiting for birth dependencies.