Ensure kubexit does not hang when watched pod is ready when started (#7)
If kubexit is started after a birth dependency has already become ready, kubexit never finds out about the ready status of the dependency.

This is because kubexit only re-evaluates the ready status on an modified event, while such an event never occurs if the birth dependency is ready before kubexit started watching.

I fixed this bug by also taking the `Added` event into consideration. This event is always emitted when kubexit starts watching the birth dependencies. If the dependencies are already ready, it will immediatly continue.